+++ src/steps/fetchFindings.ts
@@ -1,24 +1,25 @@
 import {
+  Entity,
   IntegrationStep,
   IntegrationStepExecutionContext,
 } from '@jupiterone/integration-sdk-core';
-import SnykClient from '@jupiterone/snyk-client';
 
+import { APIClient } from '../client';
 import {
-  createCVEEntities,
-  createCWEEntities,
+  createCVEEntity,
+  createCWEEntity,
   createFindingEntity,
   createFindingVulnerabilityRelationship,
   createFindingWeaknessRelationship,
   createServiceEntity,
   createServiceFindingRelationship,
-  Project,
-  SnykVulnIssue,
 } from '../converters';
 import { FindingEntity, IntegrationConfig } from '../types';
 
-type FindingsMap = {
-  [key: string]: FindingEntity;
+type EntityCache = {
+  findingEntities: { [key: string]: FindingEntity };
+  cveEntities: { [cve: string]: Entity };
+  cweEntities: { [cwe: string]: Entity };
 };
 
 const step: IntegrationStep<IntegrationConfig> = {
@@ -38,91 +39,68 @@ const step: IntegrationStep<IntegrationConfig> = {
     instance,
     logger,
   }: IntegrationStepExecutionContext<IntegrationConfig>) {
-    const config = instance.config;
-    const findingsById: FindingsMap = {};
+    const orgId = instance.config.snykOrgId;
+    const apiClient = new APIClient(logger, instance.config);
 
-    const snyk = new SnykClient(config.snykApiKey, config.snykOrgId);
-    const allProjects: Project[] = (
-      await snyk.listAllProjects(config.snykOrgId)
-    ).projects;
+    const entityCache: EntityCache = {
+      findingEntities: {},
+      cveEntities: {},
+      cweEntities: {},
+    };
 
-    logger.info(
-      {
-        projects: allProjects.length,
-      },
-      'Fetched projects',
-    );
+    const serviceEntity = createServiceEntity(orgId);
 
-    const service = createServiceEntity(config.snykOrgId);
-    await jobState.addEntity(service);
+    await apiClient.iterateProjects(async (project) => {
+      const [projectName, packageName] = project.name.split(':');
 
-    for (const project of allProjects) {
-      const name: string[] = project.name.split(':');
-      const projectName = name[0];
-      const packageName = name[1];
-
-      const listIssuesRes = (
-        await snyk.listIssues(config.snykOrgId, project.id, {})
-      ).issues;
-
-      logger.info(
-        {
-          project: {
-            id: project.id,
-          },
-          vulnerabilities: listIssuesRes.vulnerabilities.length,
-          licenses: listIssuesRes.licenses.length,
-        },
-        'Fetched project issues',
-      );
-
-      const vulnerabilities: SnykVulnIssue[] = listIssuesRes.vulnerabilities;
-      vulnerabilities.forEach((vuln: SnykVulnIssue) => {
-        vuln.type = 'vulnerability';
-      });
-
-      const licenses: SnykVulnIssue[] = listIssuesRes.licenses;
-      licenses.forEach((license: SnykVulnIssue) => {
-        license.type = 'license';
-        license.identifiers = { CVE: [], CWE: [] };
-      });
-
-      const snykIssues: SnykVulnIssue[] = vulnerabilities.concat(licenses);
-      for (const issue of snykIssues) {
+      await apiClient.iterateIssues(project, async (issue) => {
         const finding = createFindingEntity(issue);
-        if (findingsById[finding.id]) {
-          if (!findingsById[finding.id].targets.includes(projectName)) {
-            findingsById[finding.id].targets.push(projectName);
+        if (entityCache.findingEntities[finding.id]) {
+          if (
+            !entityCache.findingEntities[finding.id].targets.includes(
+              projectName,
+            )
+          ) {
+            entityCache.findingEntities[finding.id].targets.push(projectName);
           }
         } else {
           finding.targets.push(projectName);
           finding.identifiedInFile = packageName;
-          findingsById[finding.id] = finding;
+          entityCache.findingEntities[finding.id] = finding;
 
-          const cveList = createCVEEntities(finding);
-          for (const cve of cveList) {
-            await jobState.addEntity(cve);
+          for (const cve of finding.cve) {
+            let cveEntity = entityCache.cveEntities[cve];
+            if (!cveEntity) {
+              cveEntity = createCVEEntity(cve, issue.cvssScore);
+              entityCache.cveEntities[cve] = cveEntity;
+            }
             await jobState.addRelationship(
-              createFindingVulnerabilityRelationship(finding, cve),
+              createFindingVulnerabilityRelationship(finding, cveEntity),
             );
           }
 
-          const cweList = createCWEEntities(finding);
-          for (const cwe of cweList) {
-            await jobState.addEntity(cwe);
+          for (const cwe of finding.cwe) {
+            let cweEntity = entityCache.cweEntities[cwe];
+            if (!cweEntity) {
+              cweEntity = createCWEEntity(cwe);
+              entityCache.cweEntities[cwe] = cweEntity;
+            }
             await jobState.addRelationship(
-              createFindingWeaknessRelationship(finding, cwe),
+              createFindingWeaknessRelationship(finding, cweEntity),
             );
           }
 
           await jobState.addRelationship(
-            createServiceFindingRelationship(service, finding),
+            createServiceFindingRelationship(serviceEntity, finding),
           );
         }
-      }
-    }
+      });
+    });
 
-    await jobState.addEntities(Object.values(findingsById));
+    await jobState.addEntity(serviceEntity);
+    await jobState.addEntities(Object.values(entityCache.cweEntities));
+    await jobState.addEntities(Object.values(entityCache.cveEntities));
+    await jobState.addEntities(Object.values(entityCache.findingEntities));
   },
 };