+++ CHANGELOG.md
@@ -8,11 +8,18 @@ and this project adheres to
 
 ## [Unreleased]
 
+## 2.0.4 - 2020-06-26
+
+### Fixed
+
+- Removed dupliate CWE/CVE entities.
+- Handle no issues on response error.
+
 ## 2.0.3 - 2020-06-26
 
 ### Fixed
 
-- Add entities to job state 🤦🏼‍♂️
+- Add entities to job state. 🤦🏼‍♂️
 
 ## 2.0.2 - 2020-06-24